fix:extra packages parsing from file

+++ run_imagebuilder.sh
@@ -79,7 +79,7 @@ fi
 # Show a message with the list of packages in the file, replace spaces with new lines if custoḿ_packages is not empty
 if [ "$extra_packages" != "" ]; then
     dialog --title "Extra Packages" --msgbox "$(cat "$extra_packages" | tr ' ' '\n')" 20 60
-    extra_packages=$(cat "$extra_packages" | tr ' ' '\n')
+    extra_packages=$(cat "$extra_packages")
 fi
 # Show a message with the list of files in the custom path if it is not empty
 if [ "$custom_path" != "" ]; then
@@ -89,8 +89,13 @@ fi
 # Get the name of the downloaded file
 OPENWRT_FILE=$(wget -qO- $OPENWRT_URL/$SELECTED_VERSION/$WGET_PATH/ | grep -oP 'href="(openwrt-imagebuilder-[^"]+)"' | cut -d'"' -f2)
 
-# Donwload the OpenWRT imagebuilder tar file
-wget -q $OPENWRT_URL/$SELECTED_VERSION/$WGET_PATH/$OPENWRT_FILE -O builds/$OPENWRT_FILE
+# Donwload the OpenWRT imagebuilder tar file if it doesn't exist
+if [ ! -f "builds/$OPENWRT_FILE" ]; then
+    mkdir -p builds
+    wget -q $OPENWRT_URL/$SELECTED_VERSION/$WGET_PATH/$OPENWRT_FILE -O builds/$OPENWRT_FILE
+else
+    echo "File already exists, skipping download."
+fi
 
 # --------------
 # Run the command to download the imagebuilder
